production/todo_notes/todo_notes/urls.py
"""todo_notes URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/3.1/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import path, include, re_path
from django.views.generic import TemplateView
from drf_yasg import openapi
from drf_yasg.views import get_schema_view
from graphene_django.views import GraphQLView
from rest_framework import permissions
from rest_framework.authtoken import views
from rest_framework.routers import DefaultRouter
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView, TokenVerifyView

from notes.views import NoteModelViewSet, ProjectModelViewSet
from users.views import UserCustomViewSet

schema_view = get_schema_view(
    openapi.Info(
        title="ToDo Notes",
        default_version='0.1',
        description="Documentation to ToDo notes project",
        contact=openapi.Contact(email="django@gb.local"),
        license=openapi.License(name="MIT License"),
    ),
    public=True,
    permission_classes=(permissions.AllowAny,),
)

router = DefaultRouter()  # Create a router
router.register('users', UserCustomViewSet)  # Register User model view set with the router
router.register('notes', NoteModelViewSet)  # Register Note model view set with the router
router.register('projects', ProjectModelViewSet)  # Register Project model view set with the router

urlpatterns = [
    # path('', TemplateView.as_view(template_name='index.html')),

    path('admin/', admin.site.urls),
    path('api/', include(router.urls)),  # These urls are now determined automatically by the router
    path('api-token-auth/', views.obtain_auth_token),
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/token/verify/', TokenVerifyView.as_view(), name='token_verify'),

    re_path(r'^swagger(?P<format>\.json|\.yaml)$',
            schema_view.without_ui(cache_timeout=0),
            name='schema-json'),
    path('swagger/',
         schema_view.with_ui('swagger', cache_timeout=0),
         name='schema-swagger-ui'),
    path('redoc/',
         schema_view.with_ui('redoc', cache_timeout=0),
         name='schema-redoc'),
    path('graphql/', GraphQLView.as_view(graphiql=True)),
]
