production/todo_notes/requirements.txt
Django==3.1
django-cors-headers==3.8.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
graphene==2.1.9
graphene-django==2.15.0
Markdown==3.3.4
mixer==7.1.2
PyJWT==2.1.0
psycopg2-binary
